+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Remove pieces column from "Related invoice" accordion.Refs UIOR-817.
 * Update "Unopen order" confirmation modal message when pieces are deleted. Refs UIOR-827.
 * Product ID search returning random results. Refs UIOR-830.
+* Filter and identify POL by Acq unit. Refs UIOR-477.
 
 ## [3.0.0](https://github.com/folio-org/ui-orders/tree/v3.0.0) (2021-10-08)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v2.4.2...v3.0.0)